LayoutTests/imported/w3c/web-platform-tests/content-security-policy/svg/object-in-svg-foreignobject.sub-expected.txt



FAIL Should throw a securitypolicyviolation assert_equals: expected "object-src" but got "object-src 'none'"
